refactor harness.rs
