+++ packages/fuels/tests/from_token.rs
@@ -1,0 +1,215 @@
+use fuels::prelude::*;
+
+pub fn null_contract_id() -> String {
+    // a bech32 contract address that decodes to ~[0u8;32]
+    String::from("fuel1qqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqsx2mt2")
+}
+
+#[tokio::test]
+async fn create_struct_from_decoded_tokens() -> Result<(), Error> {
+    // Generates the bindings from the an ABI definition inline.
+    // The generated bindings can be accessed through `SimpleContract`.
+    abigen!(
+        SimpleContract,
+        r#"
+        {
+            "types": [
+              {
+                "typeId": 0,
+                "type": "()",
+                "components": [],
+                "typeParameters": null
+              },
+              {
+                "typeId": 1,
+                "type": "bool",
+                "components": null,
+                "typeParameters": null
+              },
+              {
+                "typeId": 2,
+                "type": "struct MyStruct",
+                "components": [
+                  {
+                    "name": "foo",
+                    "type": 3,
+                    "typeArguments": null
+                  },
+                  {
+                    "name": "bar",
+                    "type": 1,
+                    "typeArguments": null
+                  }
+                ],
+                "typeParameters": null
+              },
+              {
+                "typeId": 3,
+                "type": "u8",
+                "components": null,
+                "typeParameters": null
+              }
+            ],
+            "functions": [
+              {
+                "inputs": [
+                  {
+                    "name": "my_val",
+                    "type": 2,
+                    "typeArguments": null
+                  }
+                ],
+                "name": "takes_struct",
+                "output": {
+                  "name": "",
+                  "type": 0,
+                  "typeArguments": null
+                }
+              }
+            ]
+          }
+        "#,
+    );
+
+    // Decoded tokens
+    let u8_token = Token::U8(10);
+    let bool_token = Token::Bool(true);
+
+    // Create the struct using the decoded tokens.
+    // `struct_from_tokens` is of type `MyStruct`.
+    let struct_from_tokens = MyStruct::from_token(Token::Struct(vec![u8_token, bool_token]))?;
+
+    assert_eq!(10, struct_from_tokens.foo);
+    assert!(struct_from_tokens.bar);
+
+    let wallet = launch_provider_and_get_wallet().await;
+
+    let contract_instance = SimpleContract::new(null_contract_id(), wallet);
+
+    let call_handler = contract_instance.methods().takes_struct(struct_from_tokens);
+
+    let encoded_args = call_handler.contract_call.encoded_args.resolve(0);
+    let encoded = format!(
+        "{}{}",
+        hex::encode(call_handler.contract_call.encoded_selector),
+        hex::encode(&encoded_args)
+    );
+
+    assert_eq!("00000000cb0b2f05000000000000000a0000000000000001", encoded);
+    Ok(())
+}
+
+#[tokio::test]
+async fn create_nested_struct_from_decoded_tokens() -> Result<(), Error> {
+    // Generates the bindings from the an ABI definition inline.
+    // The generated bindings can be accessed through `SimpleContract`.
+    abigen!(
+        SimpleContract,
+        r#"
+        {
+            "types": [
+              {
+                "typeId": 0,
+                "type": "()",
+                "components": [],
+                "typeParameters": null
+              },
+              {
+                "typeId": 1,
+                "type": "bool",
+                "components": null,
+                "typeParameters": null
+              },
+              {
+                "typeId": 2,
+                "type": "struct InnerStruct",
+                "components": [
+                  {
+                    "name": "a",
+                    "type": 1,
+                    "typeArguments": null
+                  }
+                ],
+                "typeParameters": null
+              },
+              {
+                "typeId": 3,
+                "type": "struct MyNestedStruct",
+                "components": [
+                  {
+                    "name": "x",
+                    "type": 4,
+                    "typeArguments": null
+                  },
+                  {
+                    "name": "foo",
+                    "type": 2,
+                    "typeArguments": null
+                  }
+                ],
+                "typeParameters": null
+              },
+              {
+                "typeId": 4,
+                "type": "u16",
+                "components": null,
+                "typeParameters": null
+              }
+            ],
+            "functions": [
+              {
+                "inputs": [
+                  {
+                    "name": "top_value",
+                    "type": 3,
+                    "typeArguments": null
+                  }
+                ],
+                "name": "takes_nested_struct",
+                "output": {
+                  "name": "",
+                  "type": 0,
+                  "typeArguments": null
+                }
+              }
+            ]
+          }
+        "#,
+    );
+
+    // Creating just the InnerStruct is possible
+    let a = Token::Bool(true);
+    let inner_struct_token = Token::Struct(vec![a.clone()]);
+    let inner_struct_from_tokens = InnerStruct::from_token(inner_struct_token.clone())?;
+    assert!(inner_struct_from_tokens.a);
+
+    // Creating the whole nested struct `MyNestedStruct`
+    // from tokens.
+    // `x` is the token for the field `x` in `MyNestedStruct`
+    // `a` is the token for the field `a` in `InnerStruct`
+    let x = Token::U16(10);
+
+    let nested_struct_from_tokens =
+        MyNestedStruct::from_token(Token::Struct(vec![x, inner_struct_token]))?;
+
+    assert_eq!(10, nested_struct_from_tokens.x);
+    assert!(nested_struct_from_tokens.foo.a);
+
+    let wallet = launch_provider_and_get_wallet().await;
+
+    let contract_instance = SimpleContract::new(null_contract_id(), wallet);
+
+    let call_handler = contract_instance
+        .methods()
+        .takes_nested_struct(nested_struct_from_tokens);
+
+    let encoded_args = call_handler.contract_call.encoded_args.resolve(0);
+    let encoded = format!(
+        "{}{}",
+        hex::encode(call_handler.contract_call.encoded_selector),
+        hex::encode(&encoded_args)
+    );
+
+    assert_eq!("0000000088bf8a1b000000000000000a0000000000000001", encoded);
+    Ok(())
+}
